pyproject.toml
[project]
name = "atlasnode-mcp"
version = "0.1.0"
description = "Local MCP server for a SQLite hybrid vector-backed AtlasNode runtime."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mcp[cli]>=1.0,<2.0",
  "psutil>=7.0,<8.0",
  "python-dotenv>=1.0,<2.0",
  "sentence-transformers>=5.3.0,<6.0",
  "starlette>=0.27,<1.0",
  "uvicorn>=0.31,<1.0",
]

[project.scripts]
atlasnode-dashboard = "atlasnode_mcp.dashboard:main"
atlasnode-mcp = "atlasnode_mcp.server:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["atlasnode_mcp*"]


